requirements.txt
streamlit>=1.28.0,<1.40.0
groq>=0.4.0
PyPDF2>=3.0.1
sentence-transformers>=2.2.2,<3.0.0
torch>=2.0.0,<2.5.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken>=0.5.0
